+++ packages/atomic/src/components/facets/atomic-date-facet/atomic-date-facet.tsx
@@ -1,4 +1,4 @@
-import {Component, Prop, State, h, Element} from '@stencil/core';
+import {Component, Prop, State, h, Element, Host} from '@stencil/core';
 import dayjs from 'dayjs';
 import {
   DateFacet,
@@ -179,18 +179,20 @@ export class AtomicDateFacet implements InitializableComponent, BaseFacetState {
     }
 
     if (!this.facetState.hasActiveValues && this.totalNumberOfResults === 0) {
-      return;
+      return <Host class="atomic-without-values"></Host>;
     }
 
     return (
-      <BaseFacet
-        controller={new BaseFacetController(this)}
-        label={this.strings[this.label]()}
-        hasActiveValues={this.facetState.hasActiveValues}
-        clearAll={() => this.facet.deselectAll()}
-      >
-        <ul>{this.values}</ul>
-      </BaseFacet>
+      <Host class="atomic-with-values">
+        <BaseFacet
+          controller={new BaseFacetController(this)}
+          label={this.strings[this.label]()}
+          hasActiveValues={this.facetState.hasActiveValues}
+          clearAll={() => this.facet.deselectAll()}
+        >
+          <ul>{this.values}</ul>
+        </BaseFacet>
+      </Host>
     );
   }
 }