+++ packages/headless/src/api/search/trigger.ts
@@ -16,6 +16,10 @@ export interface TriggerNotify {
   content: string;
 }
 
+export function isNotifyTrigger(trigger: Trigger): trigger is TriggerNotify {
+  return trigger.type === 'notify';
+}
+
 /**
  * Redirect the user to another url
  */
@@ -24,7 +28,7 @@ export interface TriggerRedirect {
   content: string;
 }
 
-export function isTriggerRedirect(
+export function isRedirectTrigger(
   trigger: Trigger
 ): trigger is TriggerRedirect {
   return trigger.type === 'redirect';
@@ -36,6 +40,10 @@ export function isTriggerRedirect(
 export interface TriggerQuery {
   type: 'query';
   content: string;
+}
+
+export function isQueryTrigger(trigger: Trigger): trigger is TriggerQuery {
+  return trigger.type === 'query';
 }
 
 /**