+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.32.8](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.32.7...@coveo/quantic@0.32.8) (2021-07-05)
+
+
+### Bug Fixes
+
+* **quantic:** jest unit tests ([#942](https://github.com/coveo/ui-kit/issues/942)) ([1ec27f4](https://github.com/coveo/ui-kit/commit/1ec27f4fac2abc0c80ba786eca608c6bc2dc52dd))
+
+
+
+
+
+## [0.32.7](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.32.6...@coveo/quantic@0.32.7) (2021-07-05)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [0.32.6](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.32.5...@coveo/quantic@0.32.6) (2021-07-02)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [0.32.5](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.32.4...@coveo/quantic@0.32.5) (2021-07-02)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [0.32.4](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.32.3...@coveo/quantic@0.32.4) (2021-07-02)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [0.32.3](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.32.2...@coveo/quantic@0.32.3) (2021-06-30)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [0.32.2](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.32.1...@coveo/quantic@0.32.2) (2021-06-25)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
 ## [0.32.1](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.32.0...@coveo/quantic@0.32.1) (2021-06-23)
 
 **Note:** Version bump only for package @coveo/quantic